refactor theme, cleanup
