+++ app/src/main/kotlin/vadiole/unicode/ui/table/CharRow.kt
@@ -8,14 +8,16 @@ import android.view.HapticFeedbackConstants
 import android.view.MotionEvent
 import android.view.View
 import android.view.ViewConfiguration
+import vadiole.unicode.UnicodeApp.Companion.themeManager
 import vadiole.unicode.data.CodePoint
-import vadiole.unicode.ui.theme.*
+import vadiole.unicode.ui.theme.ThemeDelegate
+import vadiole.unicode.ui.theme.key_windowRipple
+import vadiole.unicode.ui.theme.key_windowTextPrimary
 import vadiole.unicode.utils.extension.dp
 import kotlin.math.floor
 
 class CharRow(
     context: Context,
-    appTheme: AppTheme,
     private val count: Int,
     private val delegate: Delegate
 ) : View(context), ThemeDelegate {
@@ -46,7 +48,7 @@ class CharRow(
     private var actionDownIndex = -1
 
     init {
-        appTheme.observe(this)
+        themeManager.observe(this)
         charPaint.textSize = charSize
         isClickable = true
         isFocusable = true
@@ -129,9 +131,9 @@ class CharRow(
 
     override fun hasOverlappingRendering() = false
 
-    override fun applyTheme(theme: Theme) {
-        charPaint.color = theme.getColor(key_windowTextPrimary)
-        ripplePaint.color = theme.getColor(key_windowRipple)
+    override fun applyTheme() {
+        charPaint.color = themeManager.getColor(key_windowTextPrimary)
+        ripplePaint.color = themeManager.getColor(key_windowRipple)
         invalidate()
     }
