+++ app/src/main/kotlin/vadiole/unicode/ui/theme/Keys.kt
@@ -1,0 +1,20 @@
+package vadiole.unicode.ui.theme
+
+const val key_dialogBackground = "dialogBackground"
+const val key_dialogSurface = "dialogSurface"
+const val key_dialogSurfacePressed = "dialogSurfacePressed"
+const val key_dialogIcon = "dialogIcon"
+const val key_dialogButton = "dialogButton"
+const val key_dialogButtonDanger = "dialogButtonDanger"
+const val key_dialogListRipple = "dialogListRipple"
+const val key_dialogDim = "dialogDim"
+const val key_windowBackground = "windowBackground"
+const val key_windowTextPrimary = "windowTextPrimary"
+const val key_windowTextSecondary = "windowTextSecondary"
+const val key_windowTextAccent = "windowTextAccent"
+const val key_windowDivider = "windowDivider"
+const val key_windowRipple = "windowRipple"
+const val key_topBarBackground = "key_toolBarBackground"
+const val key_tabBarBackground = "key_tabBarBackground"
+const val key_tabBarItem = "key_tabBarItem"
+const val key_tabBarItemSelected = "key_tabBarItemSelected"
