+++ app/src/main/kotlin/vadiole/unicode/ui/theme/ThemeDelegate.kt
@@ -3,7 +3,7 @@ package vadiole.unicode.ui.theme
 import android.view.View
 
 interface ThemeDelegate {
-    fun applyTheme(theme: Theme)
+    fun applyTheme()
     fun addOnAttachStateChangeListener(listener: View.OnAttachStateChangeListener)
     fun isAttachedToWindow(): Boolean
     fun invalidate()
